SELECT *
At first glance when you connect data and decide the steps to do 
The asterisk (*) is a wildcard character that means everything. The "SELECT *" statement is used to select all the information about data set. 
An asterisk (" * ") can be used to specify that the query should return all columns of the queried tables. SELECT is the most complex statement in SQL, with optional keywords and clauses that include: The FROM clause, which indicates the table(s) to retrieve data from. (Reference: Wikipedia)